hybrid-nodes: allow connecting a client wireless device (CPE) to a hybrid node
When a wireless device is connected as a client (i.e. like a CPE) to a remote
hybrid node (a main routing device with a "dummy" wireless AP bridge router),
the physical connection is made to the AP but the IPv4 addressing is obtained
from the main routing device.

This commit allows creating these connections, not yet removing them, with the
following restrictions:

 · the link between main routing device and wireless bridge must be of type
   "Hybrid node", which is to be added

 · the main routing device the wireless bridge is connected to must be named
   "wLan/Lan" or "wLan"